Add check for already present cloned repo folders

+++ firefox-ios/import-strings.sh
@@ -1,7 +1,11 @@
 
 echo "\n\n[*] Cloning required repo to import strings"
-rm -rf LocalizationTools
-rm -rf firefoxios-l10n
+
+if [ -d "LocalizationTools" ] || [ -d "firefoxios-l10n" ]; then
+    rm -rf LocalizationTools
+    rm -rf firefoxios-l10n
+fi
+
 git clone https://github.com/mozilla-mobile/LocalizationTools.git || exit 1
 git clone --depth 1 https://github.com/mozilla-l10n/firefoxios-l10n || exit 1
 
@@ -18,3 +22,7 @@ echo "\n\n[*] Importing Strings - takes a minute. (output in import-strings.log)
   --l10n-project-path "$PWD/../firefoxios-l10n") > import-strings.log 2>&1
 
 echo "\n\n[!] Strings have been imported. You can now create a PR."
+
+echo "\n\n[!] Clean up cloned repos"
+rm -rf LocalizationTools
+rm -rf firefoxios-l10n
